==> Batch_tools/01_Combining_Corpus_Stats.ipynb <==
import pandas as pd
import numpy as np
import os

!pwd

aapb_metadata=pd.read_csv("Tables/AAPB_ARLO_400hrs_161007.csv")


os.chdir(os.path.expanduser('~/Dropbox/aapb-hipstas'))

#
os.chdir('/Volumes/McLaughlin-6TB-1/Dropbox/aapb-hipstas/AAPB_all_durations')

corpora=['AAPB_Test_Haystack_duration_seconds.txt',\
'test_set_616_clips_duration_seconds.txt',\
'AAPB_400_hrs_labeled_duration_seconds.txt']


for filename in corpora:
    print(np.sum([float(item) for item in open(filename).read().splitlines()])/60/60)


totals=[]

for filename in [item for item in os.listdir('./') if ("_seconds.txt" in item)&(item not in corpora)]:
    totals.append(np.sum([float(item) for item in open(filename).read().splitlines()])/60/60)
    print(filename)
    
print(np.sum(totals))

aapb_metadata['File Duration']=""

aapb_metadata.iloc[2]

filenames=os.listdir(os.path.expanduser("~/Dropbox/aapb-hipstas/AAPB_all_durations"))
os.chdir(os.path.expanduser("~/Dropbox/aapb-hipstas/AAPB_all_durations"))

dur_dict={}

for filename in [fname for fname in filenames if "duration_filenames" in fname]:
    names = open(filename).read().splitlines()
    durations = open(filename.replace("duration_filenames","duration_seconds")).read().splitlines()
    if len(names)!=len(durations):
        print(filename)
    for i in range(len(names)):
        dur_dict[names[i][:-4]]=durations[i]
    
#Jim_Lehrer_duration_filenames.txt
#Jim_Lehrer_duration_seconds.txt

filename=[fname for fname in filenames if "duration_filenames" in fname][0]

dur_dict['cpb-aacip-15-sf2m61c214__barcode9544_clip1_.h264']

aapb_metadata.columns

#Adding duration for existing rows

for i in range(len(aapb_metadata)):
    row=aapb_metadata.iloc[i]
    basename=str(row['Filename'])[:-4]
    try:
        #print(str(dur_dict[basename]))
        aapb_metadata.set_value(i, 'File Duration', dur_dict[basename])
    except:
        print(basename)
    


aapb_metadata.iloc[1099]

hms=aapb_metadata.iloc[1099]['Timecode IN']

def hms_to_sec(hms):
    vals=[int(item) for item in hms.split(':')]
    return (60*60*vals[0])+(60*vals[1])+(vals[2])

hms_to_sec(hms)


for i in range(len(aapb_metadata)):
    row=aapb_metadata.iloc[i]
    #print(row['Timecode IN'])
    seconds_in=hms_to_sec(row['Timecode IN'])
    #print(row['Timecode OUT'])
    seconds_out=hms_to_sec(row['Timecode OUT'])
    try:
        aapb_metadata.set_value(i, 'Timecode IN', seconds_in)
        aapb_metadata.set_value(i, 'Timecode OUT', seconds_out)
    except:
        print(row['Filename'])
    


aapb_metadata.columns

#aapb_metadata[['AAPB Unique Identifier (GUID)', 'Type', 'Value', 'Timecode IN',
#       'Timecode OUT', 'File Duration','SonyCi ID', 'Filename']].to_csv('../AAPB_ARLO_400hrs_170412.csv',index=False)

columns=['AAPB Unique Identifier (GUID)', 'Type', 'Value', 'Timecode IN',
       'Timecode OUT', 'File Duration','SonyCi ID', 'Filename']

extended_corpus_table=pd.DataFrame(columns=['Timecode IN', 'class_no', 'tag_duration', 'class ID', 'Filename'])

speaker_dirs=os.listdir(os.path.expanduser("~/Dropbox/aapb-hipstas/Annotations_Human_Corrected"))
os.chdir(os.path.expanduser('~/Dropbox/aapb-hipstas/Annotations_Human_Corrected'))

table=[]
for speaker_dir in [item for item in speaker_dirs if '.DS_Store' not in item]:
    csvs=[item for item in os.listdir(speaker_dir) if "_corrected.csv" in item]
    for csv in csvs:
            try:
                table=pd.read_csv(speaker_dir+'/'+csv)
                table.columns=["Timecode IN","class_no","tag_duration","class ID"]
                basename=csv.replace("_corrected.csv","")
                table['Filename']=basename
                extended_corpus_table=extended_corpus_table.append(table.dropna(),ignore_index=True)
            except: pass


extended_corpus_table.tail(10)

extended_corpus_table["Timecode OUT"]=""

for i in range(len(extended_corpus_table)):
    row=extended_corpus_table.iloc[i]
    #print(row['Timecode IN'])
    seconds_in=float(row['Timecode IN'])
    #print(row['Timecode OUT'])
    tag_dur=float(row['tag_duration'])
    #print(seconds_in+tag_dur)
    extended_corpus_table.set_value(i, 'Timecode OUT', seconds_in+tag_dur)



#Adding duration for existing rows

extended_corpus_table["File Duration"]=""

for i in range(len(extended_corpus_table)):
    row=extended_corpus_table.iloc[i]
    basename=str(row['Filename'])
    try:
        #print(str(dur_dict[basename]))
        extended_corpus_table.set_value(i, 'File Duration', dur_dict[basename])
    except:
        print(basename)
    


extended_corpus_table['AAPB Unique Identifier (GUID)']=''
extended_corpus_table['Type']=''
extended_corpus_table['Value']=''
extended_corpus_table['SonyCi ID']=''


extended_corpus_table.columns

full_table=pd.concat([aapb_metadata,extended_corpus_table],ignore_index=True)

full_table=full_table[['AAPB Unique Identifier (GUID)', 'Type', 'Value','class ID', 'Timecode IN',
       'Timecode OUT', 'File Duration','SonyCi ID', 'Filename']]


full_table['Tag Duration']=''


for i in range(len(full_table)):
    row=full_table.iloc[i]
    time_in=float(row['Timecode IN'])
    time_out=float(row['Timecode OUT'])
    tag_duration=time_out-time_in
    if tag_duration<=0:
        print("Error!: "+row['Filename'])
    try:
        #print(str(dur_dict[basename]))
        full_table.set_value(i, 'Tag Duration', tag_duration)
    except:
        print(row['Filename'])
    


full_table.tail(3)




for i in range(len(full_table)):
    row=full_table.iloc[i]
    class_ID=str(row['class ID'])
    value=""
    if class_ID=="Richard Nixon":
        value="Nixon, Richard"
    elif class_ID=="Malcolm X":
        value="X, Malcolm"
    elif class_ID=="Julia Child":
        value="Child, Julia"
    elif class_ID=='Lyndon Johnson':
        value="Johnson, Lyndon"
    elif class_ID==' Julia Child':
        value="Child, Julia"
    elif class_ID=='RIchard Nixon':
        value="Nixon, Richard"
    elif class_ID=='JamesBaldwin':
        value="Baldwin, James"
    elif class_ID=='James Baldwin':
        value="Baldwin, James"
    elif class_ID=='Applause':
        value="applause"
    try:
        #print(str(dur_dict[basename]))
        if value!='':
            full_table.set_value(i, 'Value', value)
    except:
        print(row['Filename'])
    





for i in range(len(full_table)):
    row=full_table.iloc[i]
    filename=str(row['Filename'])
    new_name=""
    if filename[-4:].lower() in ['.mp4','.wav','.mp3']:
        new_name=filename[:-4]
    if new_name!='':
        full_table.set_value(i, 'Filename', new_name)

    


list(set(full_table['class ID']))

np.sum(list(full_table[full_table['class ID']=="Richard Nixon"]['Tag Duration']))/60/60

np.sum(list(full_table[full_table['Value']=="Nixon, Richard"]['Tag Duration']))/60/60

full_table=full_table[['AAPB Unique Identifier (GUID)', 'Type', 'Value','class ID', 'Timecode IN',
       'Timecode OUT', 'Tag Duration', 'File Duration','SonyCi ID', 'Filename']]


#Writing CSV

full_table.to_csv('../AAPB_ARLO_All_170412.csv',index=False)

full_table[full_table['Value']=="Nixon, Richard"]

full_table[full_table['class ID']=="Richard Nixon"]

##Stats
#aapb_metadata
#extended_corpus_table




#Total length of WGBH data
print(np.sum(list([float(item) for item in list(full_table['Tag Duration']) if item!='']))/60/60)
