==> Batch_tools/03 Metadata launcher.ipynb <==
import pandas as pd
import os

os.chdir(os.path.expanduser('~/Dropbox/aapb-hipstas'))
aapb_metadata_all=pd.read_csv("AAPB_ARLO_All_170412.csv")
aapb_metadata_all.tail(2)


filenames=list(set(list(aapb_metadata_all[aapb_metadata_all['Value']=="Baldwin, James"]['Filename'])))
filenames=[item.replace('.h264','').replace('aacip-','aacip_') for item in filenames]
filenames

pbcore_dir="/Volumes/McLaughlin-6TB-1/Dropbox/aapb-hipstas/PBCore_XML"

pbcore_filenames=os.listdir(pbcore_dir)

for item in filenames:
    if item+'.pbcore' in pbcore_filenames:
        print(item)
        os.system('open -a TextWrangler '+os.path.join(pbcore_dir,item+'.pbcore'))
