==> Batch_tools/04 YouTube Downloader.ipynb <==
import os

yt_dir="/Volumes/McLaughlin-6TB-1/Dropbox/aapb-hipstas/YouTube_corpus"

os.chdir(yt_dir)

value="King, Martin Luther"

yt_urls=open(value+'.txt').read().splitlines()
yt_urls

try: os.mkdir(value)
except: pass
os.chdir('./'+value)

for url in yt_urls:
    os.system("youtube-dl --extract-audio --audio-format wav "+url)

os.chdir(yt_dir)
