==> Batch_tools/05_Batch_applause_detection.ipynb <==
import pandas as pd
import os
import subprocess
import random

aapb_metadata_all_split=pd.read_csv("/Volumes/McLaughlin-6TB-1/Dropbox/aapb-hipstas/AAPB_ARLO_All_170414_2_sec_segs.csv")
aapb_metadata_all_split.tail(1)

## Compiling a master list of audio pathnames

pathnames=sorted([str(item) for item in list(set(list(aapb_metadata_all_split['Pathname']))) if str(item)!='nan'])

random.shuffle(pathnames)

pathnames[:3]

os.chdir('/Volumes/McLaughlin-6TB-1/Dropbox/WGBH_ARLO_Project/audio-tagging-toolkit')

import timeit
tic=timeit.default_timer()
#print(timeit.default_timer() - tic)

for pathname in pathnames:
    tic=timeit.default_timer()
    csv_name=pathname.split('/')[-1][:-10]+'.applause.csv'
    out_dir="/Volumes/McLaughlin-6TB-1/Dropbox/aapb-hipstas/Master_applause_set/"
    try:
        subprocess.call(['python2','FindApplause.py','-c','-i',pathname,'-o',out_dir+csv_name])
    except:
        print("ERROR: "+pathname)
    print(timeit.default_timer() - tic)

len(pathnames)
