==> Batch_tools/06_Batch_Feature_Extraction.ipynb <==
import pandas as pd
import os
import random
from pyAudioAnalysis import audioFeatureExtraction as aF
from pyAudioAnalysis import audioTrainTest as aT
from pyAudioAnalysis import audioBasicIO
import numpy as np
import shutil

aapb_metadata_all_split=pd.read_csv("/Volumes/McLaughlin-6TB-1/Dropbox/aapb-hipstas/AAPB_ARLO_All_170414_2_sec_segs.csv")

speakers=['Baldwin, James',
 'Buckley, William F.',
 'Bush, George HW',
 'Bush, George W.',
 'Carter, Jimmy',
 'Child, Julia',
 'Clinton, Bill',
 'Clinton, Hillary',
 'Ford, Gerald',
 'Galbraith, John Kenneth',
 'Gore, Al',
 'Johnson, Lyndon',
 'Kennedy, John F.',
 'King, Martin Luther',
 'Kissinger, Henry',
 'Lehrer, Jim',
 'Nixon, Richard',
 'Pepin, Jacques',
 'Reagan, Ronald',
 'Spanish Language',
 'Steinem, Gloria',
 'Vidal, Gore',
 'X, Malcolm']

speaker_table=aapb_metadata_all_split[aapb_metadata_all_split['Value'].isin(speakers)]



## Compiling a master list of audio pathnames
#pathnames=sorted([str(item) for item in list(set(list(aapb_metadata_all_split['Pathname']))) if str(item)!='nan'])

pathnames=sorted([str(item) for item in list(set(list(speaker_table['Pathname']))) if str(item)!='nan'])


random.shuffle(pathnames)

pathnames[:3]

for item in pathnames:
    if '16000.wav' not in item:
        print item

Fs=16000
mtWin=0.5
stWin=0.05
midTermSize = mtWin * Fs
midTermStep = 0.5
mtStep = 0.5
shortTermSize = stWin * Fs
shortTermStep = 0.05
stStep = 0.05
temp_dir="/Users/mclaugh/Desktop/temp/"


for pathname in pathnames:
    try:
        basename=pathname.split('/')[-1][:-10]
        shutil.copy(pathname,temp_dir)
        temp = aF.dirWavFeatureExtractionNoAveraging(temp_dir, mtWin, mtStep, stWin, stStep)
        np.savetxt('/Volumes/McLaughlin-6TB-1/Dropbox/aapb-hipstas/extracted_features/'+basename+'.features.csv', temp[0], delimiter=',')
        print(basename)
    except: print("ERROR: "+basename)
    temp_files = os.listdir(temp_dir)
    for filename in temp_files:
        os.remove(os.path.join(temp_dir,filename))
